latest updates from source (#28)
* Fix a malicious piston + furnace crash that causes a stackoverflow

Fixes a possibly publically known crash involving pistons and furnaces which has been observed on RetroMC and BetaLands over the last month or so.

* Improve the piston crash fix

Fixes the issue that still allowed pistons and furnaces to crash servers most probably based on the changing ID of the block. Remove the annoying printout that a person attempted to crash the server because it fired randomly at dumb times.

* Remove unnecessary debug regarding previous fix

* whoops fix that

* nope, wrong thing

im stupid

* Fix login commands being visible in console

* Bad things have happened.

* I did a bug

* kind of fix stair lighting

* Connection Pause and Login Handler cleanup

- Add [Poseidon] indicator to debug
- isActive() method for the ConnectionPause object which returns if the pause is still active.
- getCreationTime() method for the ConnectionPause object which returns the creation time of the ConnectionPause.

* Fix skeleton bow sound (#35)

* Merge the last two months of Project Poseidon into UberBukkit (#7)

* Fix a malicious piston + furnace crash that causes a stackoverflow

Fixes a possibly publically known crash involving pistons and furnaces which has been observed on RetroMC and BetaLands over the last month or so.

* Improve the piston crash fix

Fixes the issue that still allowed pistons and furnaces to crash servers most probably based on the changing ID of the block. Remove the annoying printout that a person attempted to crash the server because it fired randomly at dumb times.

* Remove unnecessary debug regarding previous fix

* Connection Pause and Login Handler cleanup

- Add [Poseidon] indicator to debug
- isActive() method for the ConnectionPause object which returns if the pause is still active.
- getCreationTime() method for the ConnectionPause object which returns the creation time of the ConnectionPause.

* Revert "Merge the last two months of Project Poseidon into UberBukkit (#7)"

This reverts commit 676da4b6d578569357962211d00417f02cc89028.

Co-authored-by: RhysB <github@johnymuffin.com>
Co-authored-by: Moresteck <moresteck@protonmail.com>
Co-authored-by: Strultz <strultzgaming@gmail.com>
Co-authored-by: Strultz <69481716+Strultz@users.noreply.github.com>
Co-authored-by: Rhys B <jetpackingwolf@gmail.com>